% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_base_class.R
\docType{class}
\name{fdic_base}
\alias{fdic_base}
\title{FDIC API Query Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fdic_base
}
\description{
A base class, from which specific queries are derived.  This class
can be used directly for a user-derived query
}
\section{Params}{

\describe{
  \item{\code{path}}{The path off of the base FDIC API URL that directs the type of query}
  \item{\code{query}}{A list of query elements passed to the web API}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$fdic_api(path, query)}}{Query using the FDIC API}
}
}

\examples{
if (curl::has_internet()) {
  x <- fdic_base$new()
  resp <- x$fdic_api("/api/institutions", 
    list(filters = "STALP:OH AND ACTIVE:1",
      fields = "ZIP,OFFDOM,CITY,COUNTY,STNAME,STALP,NAME,ACTIVE,CERT,CBSA,ASSET",
      sort_by = "ZIP",
      sort_order = "DESC",
      limit = 10,
      offset = 0,
      format = "json",
      download = "false",
      filename = "data_file"))
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
