% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_institutions_class.R
\docType{class}
\name{fdic_institutions}
\alias{fdic_institutions}
\title{FDIC API Query for Institutions}
\value{
an object of type fdic_institutions

an object containing metadata and a data.frame
}
\description{
A class to create Institutions queries to the FDIC API.  Derived
from class \code{fdic_base}.
}
\details{
Example filters
\itemize{
  \item{Filter by State name: STNAME:"West Virginia"}
  \item{Filter for any one of multiple State names: STNAME:("West Virginia","Delaware")}
  \item{Filter by last updated within an inclusive date range: DATEUPDT:["2010-01-01" TO "2010-12-31"]}
  \item{Filter for deposits over 50,000,000 (50000 thousands of dollars): DEP:[50000 TO *]}
  \item{Filter for active banks in Indiana: STALP:IA AND ACTIVE:1}
}
Example fields: \code{ZIP,OFFDOM,CITY,COUNTY,STNAME,STALP,NAME,ACTIVE,CERT,CBSA,ASSET,NETINC,DEP,DEPDOM,ROE,ROA,DATEUPDT,OFFICES}
Example sort_by: \code{OFFICES}
}
\section{Params}{

\describe{
  \item{\code{field}}{one of the available query fields returned by \code{get_available_fields}}
  \item{\code{filters}}{the filter for the bank search.}
  \item{\code{fields}}{Comma delimited list of fields to search}
  \item{\code{sort_by}}{Field name by which to sort returned data}
  \item{\code{sort_order}}{Indicator if ascending (ASC) or descending (DESC)}
  \item{\code{limit}}{The number of records to return. Default is 10 and maximum is 10,000.}
  \item{\code{offset}}{The offset of page to return}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$get_available_fields()}}{Get the fields available for queries}
  \item{\code{fdic_base$get_available_field_description(field)}}{Get the description of a field}
  \item{\code{fdic_base$setFilters(filters)}}{}
  \item{\code{fdic_base$setFields(fields)}}{}
  \item{\code{fdic_base$setSort_by(sort_by)}}{}
  \item{\code{fdic_base$setSort_order(sort_order)}}{}
  \item{\code{fdic_base$setLimit(limit)}}{}
  \item{\code{fdic_base$setOffset(offset)}}{}
}
}

\examples{
api_key_secret <- Sys.getenv("FDIC_API_KEY")
if (curl::has_internet() & api_key_secret != "") {
  x <- fdic_institutions$new(api_key_secret)
  x$get_available_fields()
  x$get_available_field_description("CBSA")
  
  x <- fdic_institutions$new(api_key_secret)
  x$setFilters("STALP:OH AND ACTIVE:1")
  x$setFields(c("ZIP","OFFDOM","CITY","COUNTY","STNAME","STALP","NAME","ACTIVE",
                "CERT","CBSA","ASSET","NETINC","DEP","DEPDOM","ROE","ROA",
                "DATEUPDT","OFFICES"))
  x$setSort_by("OFFICES")
  x$setLimit(100)
  res <- x$query_fdic()
  dim(res$data)
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
\section{Super class}{
\code{\link[fdic.api:fdic_base]{fdic.api::fdic_base}} -> \code{fdic_institutions}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fdic_institutions-new}{\code{fdic_institutions$new()}}
\item \href{#method-fdic_institutions-query_fdic}{\code{fdic_institutions$query_fdic()}}
\item \href{#method-fdic_institutions-clone}{\code{fdic_institutions$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="fdic_api"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-fdic_api'><code>fdic.api::fdic_base$fdic_api()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="get_available_field_description"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-get_available_field_description'><code>fdic.api::fdic_base$get_available_field_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="get_available_fields"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-get_available_fields'><code>fdic.api::fdic_base$get_available_fields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="parse_yaml"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-parse_yaml'><code>fdic.api::fdic_base$parse_yaml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setFields"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setFields'><code>fdic.api::fdic_base$setFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setFilters"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setFilters'><code>fdic.api::fdic_base$setFilters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setLimit"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setLimit'><code>fdic.api::fdic_base$setLimit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setOffset"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setOffset'><code>fdic.api::fdic_base$setOffset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setSort_by"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setSort_by'><code>fdic.api::fdic_base$setSort_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setSort_order"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setSort_order'><code>fdic.api::fdic_base$setSort_order()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_institutions-new"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_institutions-new}{}}}
\subsection{Method \code{new()}}{
Initialization Method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_institutions$new(api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_institutions-query_fdic"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_institutions-query_fdic}{}}}
\subsection{Method \code{query_fdic()}}{
Query FDIC API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_institutions$query_fdic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_institutions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_institutions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_institutions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
