% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_locations_class.R
\docType{class}
\name{fdic_locations}
\alias{fdic_locations}
\title{FDIC API Query for Locations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fdic_locations
}
\description{
A class to create Locations queries to the FDIC API.  Derived
from class \code{fdic_base}.
}
\details{
Example filters
\itemize{
  \item{Filter by State name: STNAME:"West Virginia"}
  \item{Filter for any one of multiple State names: STNAME:("West Virginia","Delaware")}
  \item{Filter by last updated within an inclusive date range: RUNDATE:["2015-01-01" TO "2015-01-06"]}
  \item{Filter for office number between 0 and 10: OFFNUM:[0 TO 10]}
  \item{Filter for branches in Oregon with type 11: STNAME:Oregon AND SERVTYPE:11}
}
Example fields: \code{NAME,UNINUM,SERVTYPE,RUNDATE,CITY,STNAME,ZIP,COUNTY}
Example sort_by: \code{NAME}
}
\section{Params}{

\describe{
  \item{\code{field}}{one of the available query fields returned by \code{get_available_fields}}
  \item{\code{filters}}{the filter for the bank search.}
  \item{\code{fields}}{Comma delimited list of fields to search}
  \item{\code{sort_by}}{Field name by which to sort returned data}
  \item{\code{sort_order}}{Indicator if ascending (ASC) or descending (DESC)}
  \item{\code{limit}}{The number of records to return. Default is 10 and maximum is 10,000.}
  \item{\code{offset}}{The offset of page to return}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$get_available_fields()}}{Get the fields available for queries}
  \item{\code{fdic_base$get_available_field_description(field)}}{Get the description of a field}
  \item{\code{fdic_base$setFilters(filters)}}{}
  \item{\code{fdic_base$setFields(fields)}}{}
  \item{\code{fdic_base$setSort_by(sort_by)}}{}
  \item{\code{fdic_base$setSort_order(sort_order)}}{}
  \item{\code{fdic_base$setLimit(limit)}}{}
  \item{\code{fdic_base$setOffset(offset)}}{}
}
}

\examples{
if (curl::has_internet()) {
  x <- fdic_locations$new()
  x$get_available_fields()
  x$get_available_field_description("CBSA")
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
