% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_summary_class.R
\docType{class}
\name{fdic_summary}
\alias{fdic_summary}
\title{FDIC API Query for Summary Information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fdic_summary
}
\description{
A class to create Summary queries to the FDIC API.  Derived
from class \code{fdic_base}.
}
\details{
Example filters
\itemize{
  \item{Filter by Community Banks (CB) vs. Savings Institutions (SI)}
  \item{Filter by State name: STNAME:"Virginia"}
  \item{Filter for any one of multiple State names: STNAME:("West Virginia","Delaware")}
  \item{Filter data by the year range: YEAR:["2015" TO "2017"]}
  \item{Filter for data in Alabama in 2005: STNAME:"Alabama" AND YEAR:2005}
}
Example fields: \code{STNAME,YEAR,INTINC,EINTEXP,NIM,NONII,NONIX,ELNATR,ITAXR,IGLSEC,ITAX,EXTRA,NETINC }
Example sort_by: \code{YEAR}
}
\section{Params}{

\describe{
  \item{\code{field}}{one of the available query fields returned by \code{get_available_fields}}
  \item{\code{filters}}{the filter for the bank search.}
  \item{\code{fields}}{Comma delimited list of fields to search}
  \item{\code{sort_by}}{Field name by which to sort returned data}
  \item{\code{sort_order}}{Indicator if ascending (ASC) or descending (DESC)}
  \item{\code{limit}}{The number of records to return. Default is 10 and maximum is 10,000.}
  \item{\code{offset}}{The offset of page to return}
}
}

\section{Aggregate_Params}{

\describe{
  \item{\code{agg_by}}{The field(s) by which data will be aggregated. Valid values are 'YEAR' or 'YEAR,STNAME'.}
  \item{\code{agg_term_fields}}{The field(s) for which aggregations will be counted for each unique term.}
  \item{\code{agg_sum_fields}}{The field(s) for which aggregations will be summed or aggregated.}
  \item{\code{agg_limit}}{The limit on how many aggregated results will be displayed}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$get_available_fields()}}{Get the fields available for queries}
  \item{\code{fdic_base$get_available_field_description(field)}}{Get the description of a field}
  \item{\code{fdic_base$setFilters(filters)}}{}
  \item{\code{fdic_base$setFields(fields)}}{}
  \item{\code{fdic_base$setSort_by(sort_by)}}{}
  \item{\code{fdic_base$setSort_order(sort_order)}}{}
  \item{\code{fdic_base$setLimit(limit)}}{}
  \item{\code{fdic_base$setOffset(offset)}}{}
}
}

\section{Aggregate_Usage}{

\describe{
  \item{\code{fdic_locations$setAgg_by(agg_by)}}{}
  \item{\code{fdic_locations$setAgg_term_fields(agg_term_fields)}}{}
  \item{\code{fdic_locations$setAgg_sum_fields(agg_sum_fields)}}{}
  \item{\code{fdic_locations$setAgg_limit(agg_limit)}}{}
}
}

\examples{
if (curl::has_internet()) {
  x <- fdic_summary$new()
  x$get_available_fields()
  x$get_available_field_description("INTINC")
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
