% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_branch_deposits.R
\name{get_branch_deposits}
\alias{get_branch_deposits}
\title{Get the Deposits at each Branch}
\usage{
get_branch_deposits(fdic_cert_num, year)
}
\arguments{
\item{fdic_cert_num}{The FDIC Cert number of the bank}

\item{year}{The year of the desired data \code{(>= 1994)}}
}
\value{
a data.frame of the results
}
\description{
Get the Deposits at each Branch
}
\examples{
if (curl::has_internet()) {
  get_branch_deposits(628, 2018)
}
}
