% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_institutions_class.R
\docType{class}
\name{fdic_institutions}
\alias{fdic_institutions}
\title{FDIC API Query for Institutions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fdic_institutions
}
\description{
A class to create Institutions queries to the FDIC API.  Derived
from class \code{fdic_base}.
}
\details{
Example filters
\itemize{
  \item{Filter by State name: STNAME:"West Virginia"}
  \item{Filter for any one of multiple State names: STNAME:("West Virginia","Delaware")}
  \item{Filter by last updated within an inclusive date range: DATEUPDT:["2010-01-01" TO "2010-12-31"]}
  \item{Filter for deposits over 50,000,000 (50000 thousands of dollars): DEP:[50000 TO *]}
  \item{Filter for active banks in Indiana: STALP:IA AND ACTIVE:1}
}
Example fields: \code{ZIP,OFFDOM,CITY,COUNTY,STNAME,STALP,NAME,ACTIVE,CERT,CBSA,ASSET,NETINC,DEP,DEPDOM,ROE,ROA,DATEUPDT,OFFICES}
Example sort_by: \code{OFFICES}
}
\section{Params}{

\describe{
  \item{\code{field}}{one of the available query fields returned by \code{get_available_fields}}
  \item{\code{filters}}{the filter for the bank search.}
  \item{\code{fields}}{Comma delimited list of fields to search}
  \item{\code{sort_by}}{Field name by which to sort returned data}
  \item{\code{sort_order}}{Indicator if ascending (ASC) or descending (DESC)}
  \item{\code{limit}}{The number of records to return. Default is 10 and maximum is 10,000.}
  \item{\code{offset}}{The offset of page to return}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$get_available_fields()}}{Get the fields available for queries}
  \item{\code{fdic_base$get_available_field_description(field)}}{Get the description of a field}
  \item{\code{fdic_base$setFilters(filters)}}{}
  \item{\code{fdic_base$setFields(fields)}}{}
  \item{\code{fdic_base$setSort_by(sort_by)}}{}
  \item{\code{fdic_base$setSort_order(sort_order)}}{}
  \item{\code{fdic_base$setLimit(limit)}}{}
  \item{\code{fdic_base$setOffset(offset)}}{}
}
}

\examples{
if (curl::has_internet()) {
  x <- fdic_institutions$new()
  x$get_available_fields()
  x$get_available_field_description("CBSA")
  
  x <- fdic_institutions$new()
  x$setFilters("STALP:OH AND ACTIVE:1")
  x$setFields(c("ZIP","OFFDOM","CITY","COUNTY","STNAME","STALP","NAME","ACTIVE",
                "CERT","CBSA","ASSET","NETINC","DEP","DEPDOM","ROE","ROA",
                "DATEUPDT","OFFICES"))
  x$setSort_by("OFFICES")
  x$setLimit(100)
  res <- x$query_fdic()
  dim(res$data)
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
