% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdic_summary_class.R
\docType{class}
\name{fdic_summary}
\alias{fdic_summary}
\title{FDIC API Query for Summary Information}
\value{
an object of type fdic_sod

an object containing metadata and a data.frame
}
\description{
A class to create Summary queries to the FDIC API.  Derived
from class \code{fdic_base}.
}
\details{
Example filters
\itemize{
  \item{Filter by Community Banks (CB) vs. Savings Institutions (SI)}
  \item{Filter by State name: STNAME:"Virginia"}
  \item{Filter for any one of multiple State names: STNAME:("West Virginia","Delaware")}
  \item{Filter data by the year range: YEAR:["2015" TO "2017"]}
  \item{Filter for data in Alabama in 2005: STNAME:"Alabama" AND YEAR:2005}
}
Example fields: \code{STNAME,YEAR,INTINC,EINTEXP,NIM,NONII,NONIX,ELNATR,ITAXR,IGLSEC,ITAX,EXTRA,NETINC }
Example sort_by: \code{YEAR}
}
\section{Params}{

\describe{
  \item{\code{field}}{one of the available query fields returned by \code{get_available_fields}}
  \item{\code{filters}}{the filter for the bank search.}
  \item{\code{fields}}{Comma delimited list of fields to search}
  \item{\code{sort_by}}{Field name by which to sort returned data}
  \item{\code{sort_order}}{Indicator if ascending (ASC) or descending (DESC)}
  \item{\code{limit}}{The number of records to return. Default is 10 and maximum is 10,000.}
  \item{\code{offset}}{The offset of page to return}
}
}

\section{Aggregate_Params}{

\describe{
  \item{\code{agg_by}}{The field(s) by which data will be aggregated. Valid values are 'YEAR' or 'YEAR,STNAME'.}
  \item{\code{agg_term_fields}}{The field(s) for which aggregations will be counted for each unique term.}
  \item{\code{agg_sum_fields}}{The field(s) for which aggregations will be summed or aggregated.}
  \item{\code{agg_limit}}{The limit on how many aggregated results will be displayed}
}
}

\section{Usage}{

\describe{
  \item{\code{fdic_base$get_available_fields()}}{Get the fields available for queries}
  \item{\code{fdic_base$get_available_field_description(field)}}{Get the description of a field}
  \item{\code{fdic_base$setFilters(filters)}}{}
  \item{\code{fdic_base$setFields(fields)}}{}
  \item{\code{fdic_base$setSort_by(sort_by)}}{}
  \item{\code{fdic_base$setSort_order(sort_order)}}{}
  \item{\code{fdic_base$setLimit(limit)}}{}
  \item{\code{fdic_base$setOffset(offset)}}{}
}
}

\section{Aggregate_Usage}{

\describe{
  \item{\code{fdic_locations$setAgg_by(agg_by)}}{}
  \item{\code{fdic_locations$setAgg_term_fields(agg_term_fields)}}{}
  \item{\code{fdic_locations$setAgg_sum_fields(agg_sum_fields)}}{}
  \item{\code{fdic_locations$setAgg_limit(agg_limit)}}{}
}
}

\examples{
api_key_secret <- Sys.getenv("FDIC_API_KEY")
if (curl::has_internet() & api_key_secret != "") {
  x <- fdic_summary$new(api_key_secret)
  x$get_available_fields()
  x$get_available_field_description("INTINC")
}
}
\references{
\url{https://banks.data.fdic.gov/docs/}
}
\keyword{bank}
\keyword{fdic}
\keyword{financial}
\section{Super class}{
\code{\link[fdic.api:fdic_base]{fdic.api::fdic_base}} -> \code{fdic_summary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fdic_summary-new}{\code{fdic_summary$new()}}
\item \href{#method-fdic_summary-query_fdic}{\code{fdic_summary$query_fdic()}}
\item \href{#method-fdic_summary-setAgg_by}{\code{fdic_summary$setAgg_by()}}
\item \href{#method-fdic_summary-setAgg_term_fields}{\code{fdic_summary$setAgg_term_fields()}}
\item \href{#method-fdic_summary-setAgg_sum_fields}{\code{fdic_summary$setAgg_sum_fields()}}
\item \href{#method-fdic_summary-setAgg_limit}{\code{fdic_summary$setAgg_limit()}}
\item \href{#method-fdic_summary-setMax_value}{\code{fdic_summary$setMax_value()}}
\item \href{#method-fdic_summary-setMax_value_by}{\code{fdic_summary$setMax_value_by()}}
\item \href{#method-fdic_summary-clone}{\code{fdic_summary$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="fdic_api"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-fdic_api'><code>fdic.api::fdic_base$fdic_api()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="get_available_field_description"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-get_available_field_description'><code>fdic.api::fdic_base$get_available_field_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="get_available_fields"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-get_available_fields'><code>fdic.api::fdic_base$get_available_fields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="parse_yaml"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-parse_yaml'><code>fdic.api::fdic_base$parse_yaml()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setFields"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setFields'><code>fdic.api::fdic_base$setFields()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setFilters"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setFilters'><code>fdic.api::fdic_base$setFilters()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setLimit"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setLimit'><code>fdic.api::fdic_base$setLimit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setOffset"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setOffset'><code>fdic.api::fdic_base$setOffset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setSort_by"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setSort_by'><code>fdic.api::fdic_base$setSort_by()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fdic.api" data-topic="fdic_base" data-id="setSort_order"><a href='../../fdic.api/html/fdic_base.html#method-fdic_base-setSort_order'><code>fdic.api::fdic_base$setSort_order()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-new"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-new}{}}}
\subsection{Method \code{new()}}{
Initialization Method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$new(api_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-query_fdic"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-query_fdic}{}}}
\subsection{Method \code{query_fdic()}}{
Query FDIC API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$query_fdic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setAgg_by"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setAgg_by}{}}}
\subsection{Method \code{setAgg_by()}}{
Field by which the data will be aggregated
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setAgg_by(agg_by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_by}}{field to aggregate by}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setAgg_term_fields"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setAgg_term_fields}{}}}
\subsection{Method \code{setAgg_term_fields()}}{
Fields for which aggregations will be counted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setAgg_term_fields(agg_term_fields)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_term_fields}}{field to aggregate by}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setAgg_sum_fields"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setAgg_sum_fields}{}}}
\subsection{Method \code{setAgg_sum_fields()}}{
Fields for which aggregations will be summed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setAgg_sum_fields(agg_sum_fields)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_sum_fields}}{field to aggregate by}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setAgg_limit"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setAgg_limit}{}}}
\subsection{Method \code{setAgg_limit()}}{
limit on how many aggregated fields will be displayed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setAgg_limit(agg_limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agg_limit}}{aggregation limits}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setMax_value"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setMax_value}{}}}
\subsection{Method \code{setMax_value()}}{
the field by which the max value is desired
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setMax_value(max_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_value}}{the max value field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-setMax_value_by"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-setMax_value_by}{}}}
\subsection{Method \code{setMax_value_by()}}{
the field that will be used to determine the unique records
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$setMax_value_by(max_value_by)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_value_by}}{the max value field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fdic_summary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fdic_summary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fdic_summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
